moved the changes into one file
