--- examples/flutter_gallery/lib/demo/pesto/pesto_types.dart
@@ -1,37 +1,0 @@
-// Copyright 2016 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-class Recipe {
-  const Recipe({
-    this.name,
-    this.author,
-    this.description,
-    this.imagePath,
-    this.ingredientsImagePath,
-    this.ingredients,
-    this.steps
-  });
-
-  final String name;
-  final String author;
-  final String description;
-  final String imagePath;
-  final String ingredientsImagePath;
-  final List<RecipeIngredient> ingredients;
-  final List<RecipeStep> steps;
-}
-
-class RecipeIngredient {
-  const RecipeIngredient({this.amount, this.description});
-
-  final String amount;
-  final String description;
-}
-
-class RecipeStep {
-  const RecipeStep({this.duration, this.description});
-
-  final String duration;
-  final String description;
-}
